Move GeolocateDialog to dialogs directory; style zoom buttons
